requirements.txt
ffmpeg-python
google-cloud-vision
tweepy
pytest
Flask
requests
